Broaden exceptions to catch errors when species is not found in DB
ARC tries to make parity plots to compare calculated result to RMG-database. This causes problems when the species has no data in the database, such as species containing fluorine (at the time of this commit). This can throw additional exceptions to the ones list here. ARC should not crash here if any exception is encountered, as the parity plots are not mission critical, and more importantly we relay the error for later debugging.